First attempt at adding an api.
This adds a WResource called "Calculator" at "/calc". However, you need
to call it with a valid Wt Session string, eg.
curl http://localhost:8000/calc?wtd=RrFtjA8uZQ2WVSsO"

This way works if you just want to like generate a pdf for example and
allow users to download in their browser.

Unfortunately, sessions don't last very long now, so copying and pasting
the url into curl, doesn't work.

---

The next step is to replace the call WRun call in main.cpp with our own
version, then after creating the WServer, we'll call addEntryPoint,
creating a new instance of our Calculator class, which will live as long
as the main() function lives.

In the run.sh, we need to make sure that Witty knows "/calc" and
everything under it is a public resource. eg.
--docroot=".;/resources,/calc"

or not install the app in root:

--docroot="/myapp"